scaledb/requirements old.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.6.15; python_version >= '3.6'
cffi==1.15.1
chardet==3.0.4
cryptography==2.9.2
django-cors-headers==3.1.1
django==4.1
djangorestframework==3.10.3
drf-jwt==1.13.4
idna==2.8
pycparser==2.21
pyjwt==1.7.1
requests==2.22.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
